requirements.txt
fastapi
pycryptodome
pydantic
PyJWT
pymongo
uvicorn
pydantic[email]
python-multipart
pytz
websockets
pillow
fastapi-utilities
pyyaml
openai